+++ src/main/java/http/FilePath.java
@@ -1,6 +1,6 @@
 package http;
 
-public class FileName {
+public class FilePath {
     public static final String INDEX = "/index.html";
     public static final String LOGIN_FAILED = "/user/login_failed.html";
     public static final String WRONG_ACCESS = "/wrong_access.html";
